Migrations sales

+++ database/migrations/2022_03_06_173228_create_plates_table.php
@@ -17,7 +17,7 @@ class CreatePlatesTable extends Migration
 
             $table->id();
             $table->string("name");
-            $table->decimal("basePrice");
+            $table->double("basePrice");
             $table->boolean("state");
             $table->unsignedBigInteger("idCategory");
             $table->foreign("idCategory")->references("id")->on("categories");
